+++ src/main/java/limitless/enchantments/LimitlessEnchantments.java
@@ -1,10 +1,13 @@
 package limitless.enchantments;
 
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.gamerule.v1.CustomGameRuleCategory;
 import net.fabricmc.fabric.api.gamerule.v1.GameRuleFactory;
 import net.fabricmc.fabric.api.gamerule.v1.GameRuleRegistry;
+import net.minecraft.text.Style;
+import net.minecraft.text.Text;
+import net.minecraft.util.Identifier;
 import net.minecraft.world.GameRules;
-import net.minecraft.world.GameRules.Category;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,24 +16,28 @@ public class LimitlessEnchantments implements ModInitializer {
 	public static final String MOD_ID = "limitlessenchantments";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
-	public static final GameRules.Key<GameRules.IntRule> MAX_ENCHANTMENT_LEVEL =
-	GameRuleRegistry.register("maxEnchantmentLevel", Category.MISC, GameRuleFactory.createIntRule(10, 0, 255));
+    public static final CustomGameRuleCategory ENCHANTMENTS_CATEGORY = new CustomGameRuleCategory(Identifier.of(MOD_ID, "enchantments"), Text.translatable("category.gamerule.enchantments").fillStyle(Style.EMPTY.withColor(16579668).withBold(true)));
 
+	public static final GameRules.Key<GameRules.IntRule> MAX_ENCHANTMENT_LEVEL =
+
+	GameRuleRegistry.register("maxEnchantmentLevel", ENCHANTMENTS_CATEGORY, GameRuleFactory.createIntRule(0, 0, 255));
 
     public static final GameRules.Key<GameRules.BooleanRule> NO_INCOMPATIBILITIES =
-	GameRuleRegistry.register("noIncompatibilities", Category.MISC, GameRuleFactory.createBooleanRule(true));
+	GameRuleRegistry.register("noIncompatibilities", ENCHANTMENTS_CATEGORY, GameRuleFactory.createBooleanRule(true));
+
+    public static final GameRules.Key<GameRules.BooleanRule> REBALANCED_TRADES =
+    GameRuleRegistry.register("rebalancedTrades", ENCHANTMENTS_CATEGORY, GameRuleFactory.createBooleanRule(false));
 
 	public static int MAX_ENCHANTMENT_LEVEL_INT = 255;
     public static boolean NO_INCOMPATIBILITIES_BOOLEAN = false;
+    public static boolean REBALANCED_TRADES_BOOLEAN = false;
 
 	@Override
 	public void onInitialize() {
+        
 	}
 
 	public static String convertToRoman(int number) {
-        if (number > 255)
-			number = 255;
-
         StringBuilder roman = new StringBuilder();
         int[] values = {1000, 900, 500, 400, 100, 90, 50, 40, 10, 9, 5, 4, 1};
         String[] numerals = {