+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.2
+- Added a rule to rebalance Librarian Villager trades
+- Mending now restores 2 more durabilites for every level
+- Fixed Fortune and Silk Touch incompatabilites
+- Default value of maxEnchantmentLevel is now 0
+
 # 1.0.1
 - Added `noIncompatibilities` rule
 - Default value of rule `maxEnchantmentLevel` is now 10 instead 255